backend/requirements.txt
openai>=1.30.0
fastapi>=0.111.0
pydantic-settings>=2.3.0
langchain-text-splitters>=0.2.0
langchain-ollama>=0.2.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langchain-chroma>=0.1.1
chromadb>=0.5.0
pypdf>=4.2.0
python-docx>=1.1.2
sqlalchemy>=2.0.30
cachetools>=5.3.3
slowapi>=0.1.9
python-multipart>=0.0.9
Pillow>=10.3.0
pymupdf>=1.24.0
pytesseract>=0.3.10
